requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
# Hard dependencies (must use provided packages)
accelerate @ file://localhost/build/abyss/wheels/accelerate-1.2.1-py3-none-any.whl
transformers @ file://localhost/build/abyss/wheels/transformers-4.41.0.dev0-py3-none-any.whl
tsfm_public @ file://localhost/build/abyss/wheels/tsfm_public-0.2.17-py3-none-any.whl

# Core Scientific and Data Processing Libraries
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5
tokenizers==0.19.1
urllib3==1.26.20
fsspec==2024.12.0
scikit-learn

# To be removed
h5py
nptdms
matplotlib

# Deep Learning Libraries
#--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1 # >2.3.1 does not work

# Communication and Configuration
paho-mqtt<2.0.0
aiomqtt==2.0.0  # For async MQTT publishing
pyyaml

# Hard dependencies (must use provided packages)
# ./wheels/accelerate-1.2.1-py3-none-any.whl
# ./wheels/transformers-4.41.0.dev0-py3-none-any.whl
# ./wheels/tsfm_public-0.2.17-py3-none-any.whl

# Hard dependencies (must use provided packages)
# accelerate @ file://localhost/build/abyss/src/deps/accelerate-1.2.1-py3-none-any.whl
# transformers @ file://localhost/build/abyss/src/deps/transformers-4.41.0.dev0-py3-none-any.whl
# tsfm_public @ file://localhost/build/abyss/src/deps/tsfm_public-0.2.17-py3-none-any.whl





# aiohappyeyeballs-2.6.1 aiohttp-3.11.16 aiosignal-1.3.2 async-timeout-5.0.1 datasets-3.5.0 deprecated-1.2.18 dill-0.3.8 
# frozenlist-1.5.0 fsspec-2024.12.0 multidict-6.2.0 multiprocess-0.70.16 numpy-1.26.4 propcache-0.3.1 pyarrow-19.0.1 
# regex-2024.11.6 safetensors-0.5.3 tokenizers-0.19.1 transformers-4.41.0.dev0 tsfm-public-0.2.17 urllib3-1.26.20 wrapt-1.17.2 xxhash-3.5.0 yarl-1.18.3